♻️refactor: 리턴 타입 SubTravelogueCreateRes 로 변경
