+++ src/main/java/shop/zip/travel/domain/post/subTravelogue/dto/res/SubTravelogueCreateRes.java
@@ -1,0 +1,7 @@
+package shop.zip.travel.domain.post.subTravelogue.dto.res;
+
+public record SubTravelogueCreateRes(
+    Long id
+) {
+
+}
