+++ src/main/java/shop/zip/travel/presentation/subTravelogue/SubTravelogueController.java
@@ -7,7 +7,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import shop.zip.travel.domain.post.subTravelogue.dto.SubTravelogueCreateReq;
+import shop.zip.travel.domain.post.subTravelogue.dto.req.SubTravelogueCreateReq;
+import shop.zip.travel.domain.post.subTravelogue.dto.res.SubTravelogueCreateRes;
 import shop.zip.travel.domain.post.subTravelogue.service.SubTravelogueService;
 
 @RestController
@@ -21,12 +22,13 @@ public class SubTravelogueController {
     }
 
     @PostMapping("/{travelogueId}")
-    public ResponseEntity<Long> create(
+    public ResponseEntity<SubTravelogueCreateRes> create(
         @RequestBody @Valid SubTravelogueCreateReq createReq,
         @PathVariable Long travelogueId) {
 
-        Long subTravelogueId = subTravelogueService.save(createReq, travelogueId);
-        return ResponseEntity.ok(subTravelogueId);
+        SubTravelogueCreateRes subTravelogueCreateRes = subTravelogueService.save(createReq,
+            travelogueId);
+        return ResponseEntity.ok(subTravelogueCreateRes);
     }
 
 }
